Moved most inline styles to css
