+++ frontend/.eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\index.js":"1","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\App.js":"2","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reportWebVitals.js":"3","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reducers\\index.js":"4","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\Profile.jsx":"5","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\Apartment.jsx":"6","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\Listing.jsx":"7","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reducers\\ListingDetail.js":"8","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reducers\\Home.js":"9","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\HomePage.jsx":"10","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\actions\\Home.js":"11","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\components\\HomeContent\\homeContent.jsx":"12","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\components\\Listing\\listing.jsx":"13","C:\\Users\\Mike\\code\\rentr\\frontend\\src\\actions\\ListingDetail.js":"14"},{"size":858,"mtime":1611444556014,"results":"15","hashOfConfig":"16"},{"size":763,"mtime":1612918080510,"results":"17","hashOfConfig":"16"},{"size":362,"mtime":499162500000,"results":"18","hashOfConfig":"16"},{"size":344,"mtime":1612918114040,"results":"19","hashOfConfig":"16"},{"size":545,"mtime":1611441701462,"results":"20","hashOfConfig":"16"},{"size":561,"mtime":1611441715157,"results":"21","hashOfConfig":"16"},{"size":521,"mtime":1611444897830,"results":"22","hashOfConfig":"16"},{"size":742,"mtime":1611459319162,"results":"23","hashOfConfig":"16"},{"size":984,"mtime":1612916699071,"results":"24","hashOfConfig":"16"},{"size":9929,"mtime":1612925190463,"results":"25","hashOfConfig":"16"},{"size":949,"mtime":1612916706218,"results":"26","hashOfConfig":"16"},{"size":1197,"mtime":1611461720782,"results":"27","hashOfConfig":"16"},{"size":1482,"mtime":1611461596058,"results":"28","hashOfConfig":"16"},{"size":700,"mtime":1611459267559,"results":"29","hashOfConfig":"16"},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},"pq8ptu",{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"51","messages":"52","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"53","messages":"54","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"55","messages":"56","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},{"filePath":"57","messages":"58","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"32"},"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\index.js",[],["59","60"],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\App.js",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reportWebVitals.js",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reducers\\index.js",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\Profile.jsx",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\Apartment.jsx",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\Listing.jsx",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reducers\\ListingDetail.js",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\reducers\\Home.js",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\pages\\HomePage.jsx",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\actions\\Home.js",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\components\\HomeContent\\homeContent.jsx",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\components\\Listing\\listing.jsx",[],"C:\\Users\\Mike\\code\\rentr\\frontend\\src\\actions\\ListingDetail.js",[],{"ruleId":"61","replacedBy":"62"},{"ruleId":"63","replacedBy":"64"},"no-native-reassign",["65"],"no-negated-in-lhs",["66"],"no-global-assign","no-unsafe-negation"]
