--- frontend/src/styles/Home.css
@@ -1,7 +1,0 @@
-#logo{
-  height: 50px;
-}
-
-.MuiToolbar-root.MuiToolbar-regular.MuiToolbar-gutters{
-  height: 64px;
-}
